api/requirements.txt
django==5.0.5
djangorestframework==3.15.1
Markdown==3.6
psycopg2==2.9.9
djangorestframework-simplejwt==5.3.1